+++ cmd/elephant/elephant.go
@@ -15,9 +15,9 @@ import (
 
 	"github.com/abenz1267/elephant/internal/comm"
 	"github.com/abenz1267/elephant/internal/comm/client"
-	"github.com/abenz1267/elephant/internal/common"
 	"github.com/abenz1267/elephant/internal/providers"
 	"github.com/abenz1267/elephant/internal/util"
+	"github.com/abenz1267/elephant/pkg/common"
 	"github.com/adrg/xdg"
 	"github.com/urfave/cli/v3"
 )
