+++ internal/providers/calc/setup.go
@@ -17,9 +17,9 @@ import (
 
 	_ "embed"
 
-	"github.com/abenz1267/elephant/internal/common"
 	"github.com/abenz1267/elephant/internal/providers"
 	"github.com/abenz1267/elephant/internal/util"
+	"github.com/abenz1267/elephant/pkg/common"
 	"github.com/abenz1267/elephant/pkg/pb/pb"
 )
