+++ internal/providers/desktopapplications/files.go
@@ -10,7 +10,7 @@ import (
 	"time"
 
 	"github.com/abenz1267/elephant/internal/comm/handlers"
-	"github.com/abenz1267/elephant/internal/common"
+	"github.com/abenz1267/elephant/pkg/common"
 	"github.com/adrg/xdg"
 	"github.com/charlievieth/fastwalk"
 	"github.com/fsnotify/fsnotify"
