+++ internal/providers/desktopapplications/query.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/abenz1267/elephant/internal/common"
 	"github.com/abenz1267/elephant/internal/providers"
+	"github.com/abenz1267/elephant/pkg/common"
 	"github.com/abenz1267/elephant/pkg/pb/pb"
 )
