+++ internal/providers/desktopapplications/setup.go
@@ -11,8 +11,8 @@ import (
 	"regexp"
 	"time"
 
-	"github.com/abenz1267/elephant/internal/common"
-	"github.com/abenz1267/elephant/internal/common/history"
+	"github.com/abenz1267/elephant/pkg/common"
+	"github.com/abenz1267/elephant/pkg/common/history"
 )
 
 type DesktopFile struct {
