+++ internal/providers/files/query.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/abenz1267/elephant/internal/common"
+	"github.com/abenz1267/elephant/pkg/common"
 	"github.com/abenz1267/elephant/pkg/pb/pb"
 )
