+++ internal/providers/files/setup.go
@@ -14,9 +14,9 @@ import (
 	"sync"
 	"time"
 
-	"github.com/abenz1267/elephant/internal/common"
 	"github.com/abenz1267/elephant/internal/providers"
 	"github.com/abenz1267/elephant/internal/util"
+	"github.com/abenz1267/elephant/pkg/common"
 	"github.com/djherbis/times"
 	"github.com/fsnotify/fsnotify"
 )
