+++ internal/providers/load.go
@@ -10,7 +10,7 @@ import (
 	"slices"
 	"sync"
 
-	"github.com/abenz1267/elephant/internal/common"
+	"github.com/abenz1267/elephant/pkg/common"
 	"github.com/abenz1267/elephant/pkg/pb/pb"
 	"github.com/charlievieth/fastwalk"
 )
