+++ internal/providers/menus/setup.go
@@ -11,9 +11,9 @@ import (
 	"time"
 
 	"github.com/abenz1267/elephant/internal/comm/handlers"
-	"github.com/abenz1267/elephant/internal/common"
-	"github.com/abenz1267/elephant/internal/common/history"
 	"github.com/abenz1267/elephant/internal/providers"
+	"github.com/abenz1267/elephant/pkg/common"
+	"github.com/abenz1267/elephant/pkg/common/history"
 	"github.com/abenz1267/elephant/internal/util"
 	"github.com/abenz1267/elephant/pkg/pb/pb"
 )
