+++ internal/providers/todo/setup.go
@@ -14,8 +14,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/abenz1267/elephant/internal/common"
 	"github.com/abenz1267/elephant/internal/util"
+	"github.com/abenz1267/elephant/pkg/common"
 	"github.com/abenz1267/elephant/pkg/pb/pb"
 )
