+++ internal/providers/unicode/setup.go
@@ -10,10 +10,10 @@ import (
 
 	_ "embed"
 
-	"github.com/abenz1267/elephant/internal/common"
-	"github.com/abenz1267/elephant/internal/common/history"
 	"github.com/abenz1267/elephant/internal/providers"
 	"github.com/abenz1267/elephant/internal/util"
+	"github.com/abenz1267/elephant/pkg/common"
+	"github.com/abenz1267/elephant/pkg/common/history"
 	"github.com/abenz1267/elephant/pkg/pb/pb"
 )
