+++ pkg/common/history/history.go
@@ -12,7 +12,7 @@ import (
 	"sync"
 	"time"
 
-	"github.com/abenz1267/elephant/internal/common"
+	"github.com/abenz1267/elephant/pkg/common"
 )
 
 type HistoryData struct {
