+++ pkg/common/terminal.go
